refacto components
