+++ apps/mobile/src/modules/Lightning/LNPayInfo.tsx
@@ -1,0 +1,64 @@
+import '../../../applyGlobalPolyfills';
+import React from 'react';
+import { TouchableOpacity, View } from 'react-native';
+import {  Text, TextInput } from 'react-native';
+import { useStyles } from '../../hooks';
+import stylesheet from './styles';
+
+export function LNPayInfo({
+  setIsInvoiceModalVisible,
+  setInvoiceAmount,
+  setInvoiceMemo,
+  invoiceAmount,
+  invoiceMemo,
+  isLoading,
+  generateInvoice,
+}: {
+  setIsInvoiceModalVisible: any;
+  invoiceMemo: string;
+  setInvoiceMemo: any;
+  setInvoiceAmount: any;
+  invoiceAmount: string;
+  isLoading?: boolean;
+  generateInvoice: () => void;
+}) {
+  const styles = useStyles(stylesheet);
+
+  return (
+    <View style={styles.modalOverlay}>
+      <View style={styles.modalContent}>
+        <Text style={styles.modalTitle}>Receive Satoshi</Text>
+        <View style={styles.content}>
+          <TextInput
+            placeholder="Amount (sats)"
+            value={invoiceAmount}
+            keyboardType="numeric"
+            onChangeText={setInvoiceAmount}
+            style={styles.input}
+          />
+        </View>
+        <View style={styles.content}>
+          <TextInput
+            placeholder="Notes"
+            value={invoiceMemo}
+            onChangeText={setInvoiceMemo}
+            style={styles.input}
+          />
+        </View>
+        <TouchableOpacity
+          style={[styles.button, isLoading && styles.disabledButton]}
+          onPress={generateInvoice}
+          disabled={isLoading}
+        >
+          <Text style={styles.buttonText}>{isLoading ? 'Processing...' : 'Generate Invoice'}</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.closeButton}
+          onPress={() => setIsInvoiceModalVisible(false)}
+        >
+          <Text style={styles.closeButtonText}>Close</Text>
+        </TouchableOpacity>
+      </View>
+    </View>
+  );
+}
