+++ apps/mobile/src/modules/Lightning/LNWalletInfo.tsx
@@ -1,0 +1,154 @@
+import '../../../applyGlobalPolyfills';
+import React from 'react';
+import {  Pressable, TouchableOpacity, View } from 'react-native';
+import { ActivityIndicator, Text, TextInput } from 'react-native';
+
+import { IconButton } from '../../components';
+import { useStyles } from '../../hooks';
+import stylesheet from './styles';
+
+
+
+export function LNWalletInfo({
+  connectionData,
+  balance,
+  paymentRequest,
+  preimage,
+  payInvoice,
+  handleCopyInvoice,
+  setIsZapModalVisible,
+  setIsInvoiceModalVisible,
+  isLoading,
+}: {
+  connectionData: any;
+  balance: any;
+  paymentRequest: any;
+  handleCopyInvoice: () => void;
+  preimage: any;
+  payInvoice: any;
+  setIsZapModalVisible: any;
+  setIsInvoiceModalVisible: any;
+  isLoading: boolean;
+}) {
+  const styles = useStyles(stylesheet);
+
+  if (isLoading) {
+    return <ActivityIndicator size="large" color="#0000ff" />;
+  }
+  return (
+    <View style={styles.walletcontainer}>
+      <View style={styles.infoSection}>
+        <Text style={styles.sectionTitle}>Wallet Details</Text>
+        <View style={styles.infoRow}>
+          <Text style={styles.infoLabel}>Connected to:</Text>
+          <Text style={styles.infoValue}>{connectionData?.node?.alias || 'Unknown'}</Text>
+        </View>
+        <View style={styles.infoRow}>
+          <Text style={styles.infoLabel}>Network:</Text>
+          <Text style={styles.infoValue}>{connectionData?.node?.network || 'Unknown'}</Text>
+        </View>
+        <View style={styles.infoRow}>
+          <Text style={styles.infoLabel}>Balance:</Text>
+          <Text style={styles.infoValue}>{balance ?? 'Loading...'} sats</Text>
+        </View>
+      </View>
+
+      {paymentRequest ? (
+        <View style={styles.paymentSection}>
+          <View style={styles.paymentRequest}>
+            <Text style={{ ...styles.paymentRequestLabel, fontWeight: 'bold' }}>
+              Payment Request:
+            </Text>
+
+            <Pressable>
+              <Text style={styles.paymentRequestValue} numberOfLines={1} ellipsizeMode="middle">
+                {paymentRequest ?? 'Loading...'}
+              </Text>
+              <Text>
+                <IconButton
+                  onPress={handleCopyInvoice}
+                  size={16}
+                  icon="CopyIconStack"
+                  color="primary"
+                />
+              </Text>
+            </Pressable>
+          </View>
+          <Text style={styles.paymentStatus}>
+            {preimage ? `PAID: ${preimage}` : 'Not paid yet'}
+          </Text>
+        </View>
+      ) : (
+        <Text></Text>
+      )}
+      <Pressable style={styles.button} onPress={() => setIsInvoiceModalVisible(true)}>
+        <Text style={styles.buttonText}>Receive Payment</Text>
+      </Pressable>
+
+      <View style={{ marginTop: 10, ...styles.zapSection }}>
+        <Pressable style={styles.zapButton} onPress={() => setIsZapModalVisible(true)}>
+          <Text style={styles.buttonText}>Zap a User</Text>
+        </Pressable>
+      </View>
+    </View>
+  );
+}
+
+function PayInfo({
+  setIsInvoiceModalVisible,
+  setInvoiceAmount,
+  setInvoiceMemo,
+  invoiceAmount,
+  invoiceMemo,
+  isLoading,
+  generateInvoice,
+}: {
+  setIsInvoiceModalVisible: any;
+  invoiceMemo: string;
+  setInvoiceMemo: any;
+  setInvoiceAmount: any;
+  invoiceAmount: string;
+  isLoading?: boolean;
+  generateInvoice: () => void;
+}) {
+  const styles = useStyles(stylesheet);
+
+  return (
+    <View style={styles.modalOverlay}>
+      <View style={styles.modalContent}>
+        <Text style={styles.modalTitle}>Receive Satoshi</Text>
+        <View style={styles.content}>
+          <TextInput
+            placeholder="Amount (sats)"
+            value={invoiceAmount}
+            keyboardType="numeric"
+            onChangeText={setInvoiceAmount}
+            style={styles.input}
+          />
+        </View>
+        <View style={styles.content}>
+          <TextInput
+            placeholder="Notes"
+            value={invoiceMemo}
+            onChangeText={setInvoiceMemo}
+            style={styles.input}
+          />
+        </View>
+        <TouchableOpacity
+          style={[styles.button, isLoading && styles.disabledButton]}
+          onPress={generateInvoice}
+          disabled={isLoading}
+        >
+          <Text style={styles.buttonText}>{isLoading ? 'Processing...' : 'Generate Invoice'}</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.closeButton}
+          onPress={() => setIsInvoiceModalVisible(false)}
+        >
+          <Text style={styles.closeButtonText}>Close</Text>
+        </TouchableOpacity>
+      </View>
+    </View>
+  );
+}
+
